Add SleeperProperty.isAppliedByCDK
